refactor general code for widget component/styling into separate file
